requirements.txt
flake8
appdirs==1.4.3
decorator==4.0.11
Django==1.11.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
django-registration-redux==1.6
flake8==3.3.0
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
mccabe==0.6.1
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
typing==3.6.1
wcwidth==0.1.7
